+++ client/src/components/Header.js
@@ -5,19 +5,19 @@ class Header extends Component {
     state = {users: []};
 
     componentDidMount() {
-        fetch('/heater_menu')
+        fetch('/header_menu')
             .then(res => res.json())
             .then(users => this.setState({ users }));
     }
     render() {
         return (
-            <nav className="navbar navbar-expand-lg navbar-dark bg-dark navbar-right">
-                <a className="navbar-brand" href="#">OOLL</a>
+            <nav className="navbar navbar-expand-lg navbar-light bg-light navbar-right header">
+                <a className="navbar-brand logo" href="/">OOLL</a>
                 <button className="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavDropdown" aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
                     <span className="navbar-toggler-icon"/>
                 </button>
                 <div className="collapse navbar-collapse" id="navbarNavDropdown">
-                    <ul className="navbar-nav">
+                    <ul className="navbar-nav header">
 
                         {this.state.users.map((item, index) =>
                             <li className={item.class} key={index}>
@@ -33,21 +33,3 @@ class Header extends Component {
 }
 
 export default Header;
-
-// import React from 'react'
-// import { Link } from 'react-router-dom'
-//
-// const Header = () => (
-//     <header>
-//         <nav>
-//             <ul>
-//                 <li><Link to='/'>Home</Link></li>
-//                 <li><Link to='/sign_up'>Sign up</Link></li>
-//                 <li><Link to='/test'>Test</Link></li>
-//                 <li><Link to='/caffee/100'>Caffee</Link></li>
-//             </ul>
-//         </nav>
-//     </header>
-// );
-//
-// export default Header
