+++ client/src/components/Welcome.js
@@ -4,7 +4,7 @@ const Welcome = () => (
     <div className="content">
         <h1>Все начинается с&nbsp;великолепного сайта</h1>
         <p>Дизайн и продвинутые технологии OOLL помогут создать сайт вашей мечты. Легко и бесплатно.</p>
-        <a className="link bi cta" id="mainCta" href="#" fold-name="1" target="_top">
+        <a className="link bi cta" id="mainCta" href="/" fold-name="1" target="_top">
             Создать сайт
         </a>
     </div>
